refactor: dto 패키지 컨트롤러 내부에서 외부로 변경
