+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,10 +1,9 @@
 package roomescape.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-import roomescape.controller.dto.ReservationTimeCreateRequest;
+import roomescape.dto.ReservationTimeCreateRequest;
 import roomescape.domain.ReservationTime;
 import roomescape.service.ReservationTimeService;
