+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -3,8 +3,8 @@ package roomescape.service;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.controller.dto.ReservationCreateRequest;
-import roomescape.controller.dto.ReservationTimeCreateRequest;
+import roomescape.dto.ReservationCreateRequest;
+import roomescape.dto.ReservationTimeCreateRequest;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.repository.ReservationFakeDao;
