+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -3,7 +3,7 @@ package roomescape.service;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.controller.dto.ReservationTimeCreateRequest;
+import roomescape.dto.ReservationTimeCreateRequest;
 import roomescape.domain.ReservationTime;
 import roomescape.repository.ReservationTimeFakeDao;
